move core files to root
